app/requirements.txt
django
djangorestframework
paho-mqtt
